Added Txtart.py from the MTM 2020 Grand Challenge
This is The code I submitted for the Grand Challenge for the Mastering the Mainframe Event. Its a python script that will tell a user to pick a font from a library of fonts I imported, then have them input some text and pick a color for it to be displayed as ASCII Art.